OLE_BACKEND/timetable/serializers.py
from rest_framework import serializers
from .models import Activity

class ActivitySerializer(serializers.ModelSerializer):
    class Meta:
        model = Activity
        fields = ["id", "type", "name", "date", "day", "start_time", "notes", "user"]

        