OLE_BACKEND/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==4.1.0
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
pipreqs==0.4.13
pytz==2023.3
requests==2.31.0
sqlitedict==2.1.0
sqlparse==0.4.3
urllib3==2.0.3
yarg==0.1.9